Update post model by adding author field and giving extensive permissions to access endpoints

+++ blog/views.py
@@ -3,15 +3,35 @@ from django.shortcuts import get_object_or_404
 from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework import status
-from rest_framework.decorators import APIView
-from rest_framework.permissions import IsAuthenticated
+from rest_framework.decorators import APIView, api_view, permission_classes
+from rest_framework.permissions import (
+    IsAuthenticated,
+    AllowAny,
+    IsAuthenticatedOrReadOnly,
+    IsAdminUser,
+)
 
 from drf_spectacular.utils import extend_schema
 
 from .models import Post
 from .serializers import PostSerializer, ServerDetailSerializer, ServerStatusSerializer
+from .permissions import AuthorOrReadOnlyPermission
 
 tags = ["Post"]
+
+
+@api_view(["GET", "POST"])
+@permission_classes([AllowAny])
+def homepage(request: Request):
+    if request.method == "POST":
+        data = request.data
+
+        response = {"message": "Hello World", "data": data}
+
+        return Response(data=response, status=status.HTTP_200_OK)
+    response = {"message": "Hello World"}
+
+    return Response(data=response, status=status.HTTP_200_OK)
 
 
 class ServerDetailView(APIView):
@@ -51,7 +71,7 @@ class ServerStatusView(APIView):
 
 class PostListCreateView(APIView):
     serializer_class = PostSerializer
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticatedOrReadOnly]
 
     @extend_schema(
         operation_id="list_posts",
@@ -77,7 +97,7 @@ class PostListCreateView(APIView):
         serializer_data = self.serializer_class(data=request.data)
 
         if serializer_data.is_valid():
-            serializer_data.save()
+            serializer_data.save(author=request.user)  # Assign the author
             response = {
                 "message": "Post created successfully",
                 "data": serializer_data.data,
@@ -88,16 +108,11 @@ class PostListCreateView(APIView):
 
 class PostRetrieveUpdateDeleteView(APIView):
     serializer_class = PostSerializer
-    permission_classes = [IsAuthenticated]
+    permission_classes = [AuthorOrReadOnlyPermission]
 
-    @extend_schema(
-        operation_id="retrieve_post_by_id",
-        summary="Retrieve Post by ID",
-        description="This endpoint retrieves a post by its ID",
-        tags=tags,
-    )
     def get(self, request: Request, post_id: int):
         post = get_object_or_404(Post, pk=post_id)
+        self.check_object_permissions(request, post)
         serializer_data = self.serializer_class(instance=post)
 
         response = {
@@ -106,13 +121,9 @@ class PostRetrieveUpdateDeleteView(APIView):
         }
         return Response(data=response, status=status.HTTP_200_OK)
 
-    @extend_schema(
-        summary="Update Post by ID",
-        description="This endpoint updates a post by its ID",
-        tags=tags,
-    )
     def put(self, request: Request, post_id: int):
         post = get_object_or_404(Post, pk=post_id)
+        self.check_object_permissions(request, post)
         serializer_data = self.serializer_class(instance=post, data=request.data)
 
         if serializer_data.is_valid():
@@ -124,13 +135,9 @@ class PostRetrieveUpdateDeleteView(APIView):
             return Response(response, status=status.HTTP_200_OK)
         return Response(serializer_data.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    @extend_schema(
-        summary="Delete Post by ID",
-        description="This endpoint deletes a post by its ID",
-        tags=tags,
-    )
     def delete(self, request: Request, post_id: int):
         post = get_object_or_404(Post, pk=post_id)
+        self.check_object_permissions(request, post)
         post.delete()
         response = {"message": "Post deleted successfully"}
         return Response(response, status=status.HTTP_200_OK)
